added py config module
