+++ InterfacePython/TasmanianSG.in.py
@@ -32,17 +32,15 @@ from ctypes import c_char_p, c_int, c_double, c_void_p, POINTER, cdll, create_st
 import numpy as np
 import sys
 
+import TasmanianConfig
+TasmanianInputError = TasmanianConfig.TasmanianInputError
+
 bTsgPlotting = True
 try:
     import matplotlib.pyplot as tsgPlot
 except:
     bTsgPlotting = False
     tsgPlot = []
-
-__version__ = "@Tasmanian_VERSION_MAJOR@.@Tasmanian_VERSION_MINOR@"
-__license__ = "@Tasmanian_license@"
-__author__ = "Miroslav Stoyanov"
-__git_commit_hash__ = "@Tasmanian_git_hash@"
 
 lsTsgGlobalRules = ["clenshaw-curtis", "clenshaw-curtis-zero", "chebyshev", "chebyshev-odd", "gauss-legendre", "gauss-legendre-odd", "gauss-patterson", "leja", "leja-odd", "rleja", "rleja-odd", "rleja-double2", "rleja-double4", "rleja-shifted", "rleja-shifted-even", "rleja-shifted-double", "max-lebesgue", "max-lebesgue-odd", "min-lebesgue", "min-lebesgue-odd", "min-delta", "min-delta-odd", "gauss-chebyshev1", "gauss-chebyshev1-odd", "gauss-chebyshev2", "gauss-chebyshev2-odd", "fejer2", "gauss-gegenbauer", "gauss-gegenbauer-odd", "gauss-jacobi", "gauss-jacobi-odd", "gauss-laguerre", "gauss-laguerre-odd", "gauss-hermite", "gauss-hermite-odd", "custom-tabulated"]
 lsTsgGlobalTypes = ["level", "curved", "iptotal", "ipcurved", "qptotal", "qpcurved", "hyperbolic", "iphyperbolic", "qphyperbolic", "tensor", "iptensor", "qptensor"]
@@ -51,26 +49,6 @@ lsTsgRefineTypes = ["classic", "parents", "direction", "fds", "stable"]
 lsTsgSequenceRules = ["leja", "rleja", "rleja-shifted", "max-lebesgue", "min-lebesgue", "min-delta"]
 lsTsgLocalRules = ["localp", "semi-localp", "localp-zero", "localp-boundary"]
 lsTsgAccelTypes = ["none", "cpu-blas", "gpu-default", "gpu-cublas", "gpu-cuda", "gpu-magma"]
-
-class TasmanianInputError(Exception):
-    '''Exception raised for incorret input to Tasmanian
-
-    Attributes:
-    sVariable -- string containing the variable name with incorrect value
-    sMessage -- message regarding the error
-
-    '''
-    def __init__(self, sVar, sMess):
-        self.sVariable = sVar
-        self.sMessage = sMess
-
-    def printInfo(self):
-        '''
-        prints information for the incorect function or variable
-
-        '''
-        print("Incorrect input for: {0:s}".format(self.sVariable))
-        print(self.sMessage)
 
 class TasmanianSimpleSparseMatrix:
     def __init__(self):
@@ -119,7 +97,7 @@ class TasmanianSparseGrid:
 
         '''
         if (isinstance(tasmanian_library, int)):
-            self.pLibTSG = cdll.LoadLibrary("@Tasmanian_libsparsegrid_path@")
+            self.pLibTSG = cdll.LoadLibrary(TasmanianConfig.__path_libsparsegrid__)
         elif (sys.version_info.major == 3):
             if (isinstance(tasmanian_library, str)):
                 self.pLibTSG = cdll.LoadLibrary(tasmanian_library)
